Move add button inside of day card
